requirements.txt
streamlit
langchain_core
langchain_ollama
langchain_community